[Fix] ci error
