--- presentation/presentation.properties
@@ -1,1 +1,0 @@
-API_KEY = ${API_KEY}
