+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/splash/SplashFragment.kt
@@ -28,31 +28,8 @@ class SplashFragment: Fragment() {
         (splash.drawable as Animatable).start()
 
         Handler().postDelayed({
-            if (ActivityCompat.checkSelfPermission(context!!,
-                    Manifest.permission.ACCESS_FINE_LOCATION) !=
-                PackageManager.PERMISSION_GRANTED) {
-                requestPermissions(arrayOf(Manifest.permission.ACCESS_FINE_LOCATION), 500)
-            } else {
-                startApp()
-            }
+            startApp()
         },1500)
-    }
-
-    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<String>, grantResults: IntArray) {
-        when (requestCode) {
-            500 -> {
-                if (grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED)
-                    startApp()
-                else {
-                    activity!!.finish()
-                    toast("위치정보사용을 허락 하지않아 앱을 중지합니다.")
-                }
-            }
-            else -> {
-                activity!!.finish()
-                toast("위치정보사용을 허락 하지않아 앱을 중지합니다.")
-            }
-        }
     }
 
     private fun startApp() {
