+++ presentation/src/main/res/values/google_maps_api.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 <string name="google_maps_key" translatable="false" templateMergeStrategy="preserve">
-        API_KEY
+        ${API_KEY}
     </string>
 </resources>
